requirements.txt
prometheus-client==0.13.1
pyserial==3.5
redis==4.1.2